coach-backend/requirements/local.txt
-r production.txt

Werkzeug[watchdog]==3.0.1 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
psycopg[c]==3.1.18  # https://github.com/psycopg/psycopg
watchfiles==0.21.0  # https://github.com/samuelcolvin/watchfiles

# Testing
# ------------------------------------------------------------------------------
mypy==1.7.1  # https://github.com/python/mypy
django-stubs[compatible-mypy]==4.2.7  # https://github.com/typeddjango/django-stubs
pytest==8.0.2  # https://github.com/pytest-dev/pytest
pytest-sugar==1.0.0  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs[compatible-mypy]==3.14.5  # https://github.com/typeddjango/djangorestframework-stubs

boto3
matplotlib
docker
django_cryptography
bandit
pandas==2.2.1
channels==4.2.0
daphne==4.1.2
channels_redis==4.2.1

# Documentation
sphinx==7.2.6  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2024.2.4 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==7.0.0
ruff==0.2.2  # https://github.com/astral-sh/ruff
coverage==7.4.3  # https://github.com/nedbat/coveragepy# ------------------------------------------------------------------------------

djlint==1.34.1  # https://github.com/Riverside-Healthcare/djLint
pylint-django==2.5.5
pre-commit==3.6.2  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==4.3.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.1.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.8.0  # https://github.com/pytest-dev/pytest-django
